refactor: extract approval logic into useApprovalsForApprover hook for better code organization
